cleaning up main file
